Added mocking for debug output
